client/requirements.txt
numpy==2.1.1
pandas==2.2.2
psutil==6.0.0
py-cpuinfo==9.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
websocket-client==1.8.0
